fix(serial): 修复 Connect 死锁，统一日志使用 logrus
- 修复 emitEvent 在 Connect 持有写锁时尝试获取读锁导致的死锁
- 统一所有模块使用 logrus 包级别函数，确保日志写入文件
- 修复 StopBits 类型从 int 改为 float64，支持 1.5 停止位
- 修复托盘菜单停止位多选问题
- 增加串口模块详细日志和 panic recovery

+++ pkg/config/config.go
@@ -13,7 +13,7 @@ type SerialConfig struct {
 	BaudRate int
 	DataBits int
 	Parity   string
-	StopBits int
+	StopBits float64
 }
 
 type TelnetConfig struct {
